Show an indicator while loading message details
