+++ ts/features/mvl/screens/MvlLoadingScreen.tsx
@@ -1,24 +1,11 @@
-import { View } from "native-base";
 import * as React from "react";
-import { ActivityIndicator, SafeAreaView, ScrollView } from "react-native";
-import { Body } from "../../../components/core/typography/Body";
+import { SafeAreaView } from "react-native";
 import { IOStyles } from "../../../components/core/variables/IOStyles";
-import { InfoScreenComponent } from "../../../components/infoScreen/InfoScreenComponent";
 import BaseScreenComponent from "../../../components/screens/BaseScreenComponent";
-import I18n from "../../../i18n";
 import { emptyContextualHelp } from "../../../utils/emptyContextualHelp";
+import { MessageLoading } from "../../messages/components/MessageLoading";
 
 type Props = unknown;
-
-const indicator = (
-  <ActivityIndicator
-    color={"black"}
-    size={"large"}
-    accessible={false}
-    importantForAccessibility={"no-hide-descendants"}
-    accessibilityElementsHidden={true}
-  />
-);
 
 /**
  * The loading of the MVL remote data
@@ -28,16 +15,7 @@ const indicator = (
 export const MvlLoadingScreen = (_: Props): React.ReactElement => (
   <BaseScreenComponent goBack={true} contextualHelp={emptyContextualHelp}>
     <SafeAreaView style={IOStyles.flex} testID={"MvlLoadingScreen"}>
-      <ScrollView style={[IOStyles.horizontalContentPadding]}>
-        <View spacer={true} extralarge={true} />
-        <View spacer={true} extralarge={true} />
-        <View spacer={true} extralarge={true} />
-        <InfoScreenComponent
-          image={indicator}
-          title={I18n.t("features.mvl.loading.title")}
-          body={<Body>{I18n.t("features.mvl.loading.subtitle")}</Body>}
-        />
-      </ScrollView>
+      <MessageLoading />
     </SafeAreaView>
   </BaseScreenComponent>
 );
