+++ ts/features/pn/screens/PnMessageDetailsScreen.tsx
@@ -1,7 +1,12 @@
 import * as React from "react";
+import { SafeAreaView, ScrollView } from "react-native";
+import { IOStyles } from "../../../components/core/variables/IOStyles";
+import BaseScreenComponent from "../../../components/screens/BaseScreenComponent";
 
 import { IOStackNavigationRouteProps } from "../../../navigation/params/AppParamsList";
 import { UIMessageId } from "../../../store/reducers/entities/messages/types";
+import { emptyContextualHelp } from "../../../utils/emptyContextualHelp";
+import { MessageLoading } from "../../messages/components/MessageLoading";
 import { PnParamsList } from "../navigation/params";
 
 export type PnMessageDetailsScreenNavigationParams = Readonly<{
@@ -9,5 +14,11 @@ export type PnMessageDetailsScreenNavigationParams = Readonly<{
 }>;
 
 export const PnMessageDetailsScreen = (
-  props: IOStackNavigationRouteProps<PnParamsList, "PN_DETAILS">
-): React.ReactElement => <></>;
+  props: IOStackNavigationRouteProps<PnParamsList, "PN_ROUTES_MESSAGE_DETAILS">
+): React.ReactElement => (
+  <BaseScreenComponent goBack={true} contextualHelp={emptyContextualHelp}>
+    <SafeAreaView style={IOStyles.flex}>
+      <MessageLoading />
+    </SafeAreaView>
+  </BaseScreenComponent>
+);
